requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
accelerate>=0.24.0
evaluate>=0.4.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0,<3.0.0
hydra-core>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
